pyproject.toml
[project]
name = "6img-to-3d"
version = "1.0.0"
description = "Few-Image Large-Scale Outdoor Novel View Synthesis"
readme = "README.md"
license = {text = "BSD-3-Clause"}
requires-python = ">=3.10"
dependencies = [
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "ninja",
    "setuptools<81",  # mmcv build needs pkg_resources, removed in setuptools 82
    "mmengine>=0.10.0",
    # mmcv must be built from source with CUDA toolkit — see README
    "mmsegmentation>=1.2.0",
    "pyyaml>=6.0",
    "imageio>=2.33",
    "imageio-ffmpeg>=0.4.9",
    "lpips>=0.1.4",
    "pytorch-msssim>=1.0.0",
    "kornia>=0.7.0",
    "seaborn>=0.13",
    "crc32c",
    "pandas",
    "tensorboardX",
    "matplotlib",
    "opencv-python",
    "scikit-image",
    "tqdm",
    "torchmetrics",
]

[dependency-groups]
dev = [
    "ruff",
    "pre-commit",
    "pytest",
]
jupyter = [
    "jupyterlab",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = [
    "builder",
    "config",
    "dataloader",
    "triplane_decoder",
    "triplane_encoder",
    "utils",
]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "W"]
ignore = [
    "E402",   # module level import not at top of file (common pattern in ML code)
    "E501",   # line too long - handled by formatter
    "E722",   # bare except (existing research code pattern)
    "E731",   # lambda assignment (existing research code pattern)
    "E741",   # ambiguous variable name (common in math-heavy code)
    "F403",   # star imports (used for registry side-effects)
    "F405",   # may be undefined from star imports
    "F841",   # local variable assigned but never used (common in research code)
]

[tool.ruff.lint.isort]
known-first-party = ["builder", "config", "dataloader", "triplane_decoder", "triplane_encoder", "utils"]

[tool.uv.sources]
torch = { index = "pytorch-cu126" }
torchvision = { index = "pytorch-cu126" }

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
